Split protocol query codecs
